Allow Content-Type in request header sent to backend. Hook update delete button to actually delete security
